backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
httpx>=0.28.0
polyline>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
